Use floats instead of doubles
Idea stolen from:

  https://github.com/hansonw/fuzzy-native/commit/cbed93b253a

This yields a surprisingly large speed boost, more than the ~10%
mentioned in that commit. It looks like almost 22%, and the loss of
precision seems harmless enough (at least no tests failed).

Note the switch from `DBL_MAX` as a sentinel value to `FLT_MAX`. Without
this, naturally, all hell breaks loose.

  Before (best of three runs):

                            user     system      total        real
    pathological        0.270000   0.010000   0.280000 (  0.268869)
    command-t           1.410000   0.000000   1.410000 (  1.420606)
    chromium (subset)   1.140000   0.160000   1.300000 (  0.250876)
    chromium (whole)    4.880000   0.020000   4.900000 (  0.679532)

  After (best of three runs):

                            user     system      total        real
    pathological        0.200000   0.000000   0.200000 (  0.197547)
    command-t           1.250000   0.010000   1.260000 (  1.265484)
    chromium (subset)   0.960000   0.130000   1.090000 (  0.217825)
    chromium (whole)    3.850000   0.020000   3.870000 (  0.533365)

Committing for historical posterity, but will have to revert.

+++ ruby/command-t/match.h
@@ -3,7 +3,7 @@
 
 #include <ruby.h>
 
-extern double calculate_match(
+extern float calculate_match(
     VALUE str,
     VALUE needle,
     VALUE case_sensitive,
